requirements.txt
anyio==4.10.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
Brotli==1.1.0
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.1
cssselect2==0.8.0
daphne==4.2.1
Django==5.2.5
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.0
fonttools==4.59.1
h11==0.16.0
httptools==0.6.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.1
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyphen==0.17.2
python-decouple==3.8
python-dotenv==1.1.1
pywin32==311
PyYAML==6.0.2
redis==6.4.0
reportlab==4.4.3
service-identity==24.2.0
sniffio==1.3.1
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
watchfiles==1.1.0
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
whitenoise==6.9.0
zope.interface==8.0.1
zopfli==0.2.3.post1
